pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "haniel"
version = "0.1.0"
description = "Configuration-based, intentionally ignorant service runner"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Dorothy", email = "dorothy@example.com" }
]
keywords = ["service", "runner", "process", "supervisor", "git", "deployment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
haniel = "haniel.cli:main"

[project.urls]
Homepage = "https://github.com/eiaserinnys/Haniel"
Repository = "https://github.com/eiaserinnys/Haniel"
Issues = "https://github.com/eiaserinnys/Haniel/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/haniel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/haniel"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
